requirements.txt
asgiref==3.6.0
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
